refactor(vm): move opcode dispatcher into vm_exec module
